[REFACTOR] 9조 머지 후 일부 변경

+++ Airbnb_backend_9/src/main/java/project/airbnb_backend_9/domain/Review.java
@@ -50,7 +50,7 @@ public class Review {
     }
 
     @Builder
-    public Review(Long rating, String comment, Users users, Accommodation accommodation){
+    public Review(int rating, String comment, Users users, Accommodation accommodation){
         this.rating=rating;
         this.comment=comment;
         this.users=users;
